Add color-changing functionality
Replaced the QLabels that display the states of the Hall Effect sensors
with QLineEdit objects. This is because they can use the QPalette::Base
property to have a background color, which is set to red or green based
on the state of the relevant sensor.

The color-changing functionality follows these basic steps:

1. Create a QPalette from the current palette of the desired QLineEdit
2. Change the QPalette:Base property to the desired color
3. Set the QLineEdit's palette to be the modified QPalette