OBCG-2: Add prototype board selection functionality
Adds a QComboBox widget to the right hand side of the screen below the
Hall Effect sensor readouts. The combo box is preloaded with the
strings "Primary" and "Backup 1" to indicate which OBC is being used.
Default value is "Primary".

On user change of the value of the combo box, the new index is stored to
the "board_selection" int to be passed on to parsePressureMessage, from
which it is then passed on to tc_conv.

tc_conv runs the board_selection through a switch to choose between two
different sets of arrays of coefficients for thermocouple conversion.