OBCG-2: Make PT conversions depend on board selection
This commit removes board switching functionality from tc_conv, as the
thermocouple conversion equation does not change from board to board and
is a static function of the thermocouple itself.

The PT conversions now use the same switching system as described in
commit 58c71c3a, though limited to only the bias defines as those are
the only values that should change between boards.

This commit also removes the old parseMessage function that has been
deprecated for some time in favor of parsePressureMessage, since it
would not be worth the time to also convert parseMessage for the
board-switching functionality.