OBCG-17: Fix bug where liveplots are unusable
Somewhere in the range of March or April in 2019, a change was made to
the MainWindow::getData function where it started retrieving the
timestamp for datapackets from the timestamp sent by the OBC, rather
than generating a timestamp at reciept. However, when this change was
made, the serial::parsePressureMessage function was not transferring
timestamp values from the unparsed to parsed data structs. As such, the
liveplots became unusable as the timestamp value they were accessing was
the uninitialized memory in the struct field.

Fix: Added statement to transfer timestamp data to parsed data struct